renamed search.py to search_dialog.py
