+++ openpype/tools/settings/settings/search_dialog.py
